requirements.txt
matplotlib
numpy
pillow
dwave-tabu